requirements.txt
asgiref==3.11.1
Django==6.0.2
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
redis==7.2.0
sqlparse==0.5.5
tzdata==2025.3
